Remove rpi3-fta device profile and mate-fta edition profile
Darksky mentioned that these editions are no longer in use and was created for a very special usecase. So remove them from the official profiles.

Users of these profiles can maintain a local custom version if they desire, but some of the packages might go missing from the repo at some point.

Signed-off-by: Dan Johansen <strit@manjaro.org>