remove qt5-es2-wayland
Rodney from KDE suggested to remove qt5-es2-wayland package as it
is not different at all to regulat qt5-wayland.
And since nothing in our repository depends directly on it,
this should be easy to do.

But lets start with just not installing it by default on the
Dev images for testing, this will make it install regular qt5-wayland
instead.

Signed-off-by: Dan Johansen <strit@manjaro.org>