[plasma-mobile] cleanup installed applications
As proposed by #11, this is the beginning of an effort
to clean up the installed applications in our default
Plasma Mobile images. As these images is the ones people
see when they buy the phones it should only contain
what makes sense for a new user.

Any application can still be installed on the system
by normal means, so if anyway is missing any of these,
they can just install them back.

I have not removed Calindori yet, as Kalendar will pull
in all of Akonadi (over 200 MB in dependencies), so that
needs to still be discussed.

Signed-off-by: Dan Johansen <strit@manjaro.org>