[plasma-mobile-dev] add openssl-1.1
Since the update to OpenSLL 3, some things still look for openssl 1.1
library. So install the compat package for that.

Not added to regular plasma mobile profile, because that's also used by
the Manjaro ARM Installer, which targets stable branch, so it would fail
to install as openssl 3 is only in unstable at this time.

Signed-off-by: Dan Johansen <strit@manjaro.org>