De-duplicate all profiles
Clean up all the profiles to take the shared profile into account.
This means a lot of packages was removed from the profiles, but they are
all present in the "shared" one.

This update does require Manjaro ARM Tools 2.11.2 or Manjaro ARM
Installer 1.5.3. The previous ones won't work with these profiles
anymore.

Signed-off-by: Dan Johansen <strit@manjaro.org>